restructure button components
